Plgdata proxy
Base proxy class externalized. It is used by both Rimrock and Plgdata
proxies.

All request to `/plgdata` will be redirected into Plgdata service. You
can defined Plgdata url in `config/application.yml`.
