+++ app/controllers/plgdata_controller.rb
@@ -1,0 +1,11 @@
+require 'proxy/plgdata'
+
+class PlgdataController < ApplicationController
+  include Proxyable
+
+  protected
+
+  def proxy_class
+    Proxy::Plgdata
+  end
+end
