+++ lib/proxy/plgdata.rb
@@ -1,0 +1,17 @@
+require 'resolv'
+require 'base64'
+require 'proxy/base'
+
+module Proxy
+  class Plgdata < Proxy::Base
+    def initialize(proxy)
+      super(proxy, Rails.configuration.constants['plgdata'])
+    end
+
+    protected
+
+    def path(env)
+      env['PATH_INFO'].match(/\/plgdata(.*)/)[1]
+    end
+  end
+end
